moved punchouts functions state stuff into its own state
