+++ src/app/extensions/punchout/store/punchout-functions/index.ts
@@ -1,0 +1,2 @@
+// API to access ngrx punchoutFunctions state
+export * from './punchout-functions.actions';
