+++ src/app/extensions/punchout/store/punchout-store.module.ts
@@ -5,13 +5,16 @@ import { pick } from 'lodash-es';
 
 import { resetOnLogoutMeta } from 'ish-core/utils/meta-reducers';
 
+import { PunchoutFunctionsEffects } from './punchout-functions/punchout-functions.effects';
 import { PunchoutState } from './punchout-store';
 import { PunchoutUsersEffects } from './punchout-users/punchout-users.effects';
 import { punchoutUsersReducer } from './punchout-users/punchout-users.reducer';
 
-const punchoutReducers: ActionReducerMap<PunchoutState> = { punchoutUsers: punchoutUsersReducer };
+const punchoutReducers: ActionReducerMap<PunchoutState> = {
+  punchoutUsers: punchoutUsersReducer,
+};
 
-const punchoutEffects = [PunchoutUsersEffects];
+const punchoutEffects = [PunchoutUsersEffects, PunchoutFunctionsEffects];
 
 const metaReducers = [resetOnLogoutMeta];
